+++ eiam-20211201/ChangeLog.md
@@ -1,3 +1,9 @@
+2026-02-28 Version: 2.13.2
+- Update API GetPasswordComplexityConfiguration: add response parameters Body.PasswordComplexityConfiguration.DisabledWeakPasswordLogin.
+- Update API GetPasswordComplexityConfiguration: add response parameters Body.PasswordComplexityConfiguration.DisabledWeakPasswordLoginStartedAt.
+- Update API SetPasswordComplexityConfiguration: add request parameters DisabledWeakPasswordLogin.
+
+
 2026-02-27 Version: 2.13.1
 - Generated python 2021-12-01 for Eiam.